pyproject.toml
[tool.poetry]
name = "thesis-test-task"
version = "0.1.0"
description = "Тестовое задание для Thesis"
authors = ["OlegBugaichuk <olegussek@gmail.com>"]
readme = "README.md"
packages = [{include = "thesis_test_task"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"
drf-yasg = "^1.21.5"
pytest = "^7.2.2"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
